requirements.txt
Flask==3.0.2
Pillow==11.2.1
torch==2.7.0+cpu
torchvision==0.22.0+cpu
pathlib==1.0.1
tqdm==4.66.2 